printf format fix.
(Logical change 1.507)

+++ ntfsprogs/ntfsclone.c
@@ -731,9 +731,9 @@ static void walk_runs(struct ntfs_walk_cluster *walk)
 		for (j = 0; j < lcn_length; j++) {
 			u64 k = (u64)lcn + j;
 			if (ntfs_bit_get_and_set(lcn_bitmap.bm, k, 1))
-				err_exit("Cluster %lu referenced twice!\n"
+				err_exit("Cluster %llu referenced twice!\n"
 					 "You didn't shutdown your Windows"
-					 "properly?\n", k);
+					 "properly?\n", (unsigned long long)k);
 		}
 
 		walk->image->inuse += lcn_length;
